Updated Racetime.gg Information
Appends the end of the raceroom with the permalink. At some point it should also Prefix it with the style of race. 
Examples would be 
Standard : Permalink
Spoiler : Permalink
Fun : Permalink
Further Changes to !reset would be required, else we would get "Example Title - Permalink - Other Permalink"
Potentially could use str.split("-") on the title and removing the last element?